modularized dashboard
