+++ src/components/Dashboard/CallsList.tsx
@@ -1,0 +1,41 @@
+import { Card, Typography } from "antd";
+import { push } from "connected-react-router";
+import React from "react";
+import { useTranslation } from "react-i18next";
+import { Call } from "src/types/Call";
+import CallItem from "./CallItem";
+
+interface Props {
+  calls: Call[];
+}
+
+const CallsList: React.FC<Props> = ({ calls }: Props) => {
+  const { t } = useTranslation("dashboard");
+
+  return (
+    <Card
+      title={t("call.title")}
+      extra={
+        <Typography.Link onClick={() => push("/")}>
+          {t("call.seeAll")}
+        </Typography.Link>
+      }
+      style={{ borderRadius: 8 }}
+    >
+      <div style={{ paddingBottom: 16 }}>
+        <Typography.Text>
+          {calls.length === 0
+            ? t("call.noCalls")
+            : `${calls.length} ${t("call.upcomingCalls")}${
+                calls.length > 1 ? "s" : ""
+              }`}
+        </Typography.Text>
+      </div>
+      {calls.map((call) => (
+        <CallItem call={call} key={`callItem-${call.id}`} />
+      ))}
+    </Card>
+  );
+};
+
+export default CallsList;
